Migrate /blog/[slug]/edit/api/upload-image endpoint to Server Actions
